chapter-18-capstone-finance/requirements.txt
-e openai>=1.0
pandas>=2.0
python-dotenv>=1.0
